Refactor testclient into pkg
